refactor(mcp): split DCode lifecycle responsibilities
