+++ src/lib/actions/sandbox/mcp-bridge-adapter-deepagents-command.ts
@@ -1,0 +1,28 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import type { McpBridgeEntry } from "../../state/registry";
+import type { AdapterMutationOptions } from "./mcp-bridge-adapter-inspection";
+import { McpBridgeError } from "./mcp-bridge-contracts";
+import { redactBridgeSecretsForDisplay } from "./mcp-bridge-output";
+import { executeSandboxCommand } from "./process-recovery";
+
+export function runDeepAgentsAdapterCommand(
+  sandboxName: string,
+  entry: Pick<McpBridgeEntry, "env">,
+  command: string,
+  failureMessage: string,
+  options: AdapterMutationOptions = {},
+): string {
+  const result = executeSandboxCommand(sandboxName, command);
+  const output = redactBridgeSecretsForDisplay(
+    [result?.stdout, result?.stderr].filter(Boolean).join("\n").trim(),
+    entry,
+    options.envValues ?? {},
+  );
+  if (!result || result.status !== 0) {
+    if (options.bestEffort) return "";
+    throw new McpBridgeError(output || failureMessage);
+  }
+  return result.stdout;
+}
