+++ src/lib/actions/sandbox/mcp-bridge-adapter-deepagents-legacy.ts
@@ -1,0 +1,193 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import type { McpBridgeEntry } from "../../state/registry";
+import {
+  DEEPAGENTS_MCP_CONFIG_PATH,
+  deepAgentsManagedServerConfig,
+  pythonJsonLiteral,
+} from "./mcp-bridge-adapter-status";
+
+// Source-of-truth review for the legacy compatibility boundary:
+// invalidState: a v1 sandbox keeps NemoClaw's owned server inside the mutable,
+// user-shared .mcp.json file, while current images use a dedicated projection.
+// sourceBoundary: the surviving v1 Deep Agents runtime selects the legacy path;
+// the host registry remains authoritative for the exact entry NemoClaw owns.
+// whyNotSourceFix: replacing the image before teardown would strand its provider
+// and policy, so old images must be scrubbed and rolled back in their own format.
+// regressionTest: focused legacy teardown, rollback, drift, duplicate-key, mode,
+// and runtime-generation suites execute the rendered helper against real files.
+// removalCondition: delete this compatibility module after supported releases can
+// no longer contain registry-owned v1 entries and the migration window has ended.
+export const DEEPAGENTS_LEGACY_MCP_CONFIG_PATH = "/sandbox/.deepagents/.mcp.json";
+
+export const DEEPAGENTS_LEGACY_CONFIG_HELPERS = [
+  "LEGACY_MCP_MAX_BYTES = 262144",
+  "def reject_duplicate_keys(pairs):",
+  "    result = {}",
+  "    for key, value in pairs:",
+  "        if key in result:",
+  "            raise ValueError(f'duplicate JSON key: {key}')",
+  "        result[key] = value",
+  "    return result",
+  "def reject_non_json_constant(value):",
+  "    raise ValueError(f'non-JSON numeric constant: {value}')",
+  "def legacy_fingerprint(metadata):",
+  "    return (metadata.st_dev, metadata.st_ino, metadata.st_size, metadata.st_mtime_ns, metadata.st_ctime_ns, metadata.st_mode, metadata.st_nlink, metadata.st_uid)",
+  "def read_legacy_config(path):",
+  "    flags = os.O_RDONLY | os.O_CLOEXEC | os.O_NONBLOCK | os.O_NOFOLLOW",
+  "    descriptor = os.open(path, flags)",
+  "    try:",
+  "        before = os.fstat(descriptor)",
+  "        linked = os.stat(path, follow_symlinks=False)",
+  "        safe = (stat.S_ISREG(before.st_mode) and before.st_uid == os.getuid() and stat.S_IMODE(before.st_mode) == 0o600 and before.st_nlink == 1 and (before.st_dev, before.st_ino) == (linked.st_dev, linked.st_ino))",
+  "        if not safe:",
+  "            raise ValueError('legacy MCP config has unsafe ownership, mode, type, or links')",
+  "        if before.st_size <= 0 or before.st_size > LEGACY_MCP_MAX_BYTES:",
+  "            raise ValueError('legacy MCP config has invalid size')",
+  "        chunks = []",
+  "        remaining = before.st_size",
+  "        while remaining:",
+  "            chunk = os.read(descriptor, remaining)",
+  "            if not chunk:",
+  "                break",
+  "            chunks.append(chunk)",
+  "            remaining -= len(chunk)",
+  "        after = os.fstat(descriptor)",
+  "        linked_after = os.stat(path, follow_symlinks=False)",
+  "        stable = (legacy_fingerprint(before) == legacy_fingerprint(after) and legacy_fingerprint(after) == legacy_fingerprint(linked_after))",
+  "        if remaining or not stable:",
+  "            raise ValueError('legacy MCP config changed while reading')",
+  "    finally:",
+  "        os.close(descriptor)",
+  "    raw = b''.join(chunks).decode('utf-8')",
+  "    data = json.loads(raw, object_pairs_hook=reject_duplicate_keys, parse_constant=reject_non_json_constant)",
+  "    return data, legacy_fingerprint(before)",
+  "def assert_legacy_source_stable(path, identity):",
+  "    if identity is None:",
+  "        if os.path.lexists(path):",
+  "            raise ValueError('legacy MCP config appeared during mutation')",
+  "        return",
+  "    current = os.stat(path, follow_symlinks=False)",
+  "    safe = (stat.S_ISREG(current.st_mode) and current.st_uid == os.getuid() and stat.S_IMODE(current.st_mode) == 0o600 and current.st_nlink == 1 and legacy_fingerprint(current) == identity)",
+  "    if not safe:",
+  "        raise ValueError('legacy MCP config changed before mutation')",
+];
+
+export function buildDeepAgentsMcpRollbackRegisterCommand(
+  entry: McpBridgeEntry,
+  expectedServers: Record<string, Record<string, unknown>>,
+): string {
+  const payload = {
+    server: entry.server,
+    expected: deepAgentsManagedServerConfig(entry),
+    expectedServers,
+  };
+  return [
+    "/opt/venv/bin/python3 -I - <<'PY'",
+    "import json, os, pathlib, stat, sys, tempfile",
+    `payload = json.loads(${pythonJsonLiteral(payload)})`,
+    `managed_path = pathlib.Path(${JSON.stringify(DEEPAGENTS_MCP_CONFIG_PATH)})`,
+    `legacy_path = pathlib.Path(${JSON.stringify(DEEPAGENTS_LEGACY_MCP_CONFIG_PATH)})`,
+    ...DEEPAGENTS_LEGACY_CONFIG_HELPERS,
+    `runtime_kind = "auto"  # NEMOCLAW_DEEPAGENTS_RUNTIME_TEST_ANCHOR`,
+    "if runtime_kind == 'auto':",
+    "    runtime_kind = 'unknown'",
+    "    try:",
+    "        from deepagents_code import _nemoclaw_managed as managed",
+    "        runtime_path = str(getattr(managed, '_MCP_CONFIG_FILE', ''))",
+    "        if runtime_path == str(managed_path):",
+    "            runtime_kind = 'v2'",
+    "        elif runtime_path == str(legacy_path):",
+    "            runtime_kind = 'legacy'",
+    "    except Exception:",
+    "        pass",
+    "if runtime_kind not in ('v2', 'legacy'):",
+    "    print('Could not identify the managed Deep Agents MCP runtime; refusing rollback', file=sys.stderr)",
+    "    raise SystemExit(2)",
+    "is_v2 = runtime_kind == 'v2'",
+    "if is_v2 and len(payload['expectedServers']) > 64:",
+    "    print('Managed MCP v2 supports at most 64 servers', file=sys.stderr)",
+    "    raise SystemExit(2)",
+    "config_path = managed_path if is_v2 else legacy_path",
+    "data = {}",
+    "legacy_identity = None",
+    "if os.path.lexists(config_path):",
+    "    if config_path.is_symlink():",
+    "        print(f'Refusing to restore managed MCP state through symlink {config_path}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    try:",
+    "        if is_v2:",
+    "            data = json.loads(config_path.read_text(encoding='utf-8') or '{}')",
+    "        else:",
+    "            data, legacy_identity = read_legacy_config(config_path)",
+    "    except (OSError, UnicodeDecodeError, ValueError) as exc:",
+    "        print(f'Invalid managed MCP rollback state at {config_path}: {exc}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "if not isinstance(data, dict):",
+    "    print(f'Invalid managed MCP rollback state at {config_path}: expected object', file=sys.stderr)",
+    "    raise SystemExit(2)",
+    "if is_v2:",
+    "    if data and set(data) != {'mcpServers'}:",
+    "        print(f'Invalid managed MCP v2 projection at {config_path}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    servers = data.get('mcpServers', {})",
+    "    if not isinstance(servers, dict):",
+    "        print(f'Invalid managed MCP v2 server map at {config_path}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    if any(payload['expectedServers'].get(name) != current for name, current in servers.items()):",
+    "        print(f'Refusing to overwrite drifted managed MCP v2 projection at {config_path}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    data = {'mcpServers': payload['expectedServers']}",
+    "else:",
+    "    servers = data.setdefault('mcpServers', {})",
+    "    if not isinstance(servers, dict):",
+    "        print(f'Refusing to overwrite mixed legacy MCP state at {config_path}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    current = servers.get(payload['server'])",
+    "    if payload['server'] in servers and current != payload['expected']:",
+    "        print(f'Refusing to overwrite user-owned legacy MCP server at {config_path}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    servers[payload['server']] = payload['expected']",
+    "config_path.parent.mkdir(parents=True, exist_ok=True)",
+    "tmp_fd, tmp_name = tempfile.mkstemp(prefix='.nemoclaw-mcp.', dir=config_path.parent)",
+    "try:",
+    "    os.fchmod(tmp_fd, 0o600)",
+    "    with os.fdopen(tmp_fd, 'w', encoding='utf-8') as tmp_file:",
+    "        json.dump(data, tmp_file, indent=2, sort_keys=True)",
+    "        tmp_file.write('\\n')",
+    "        tmp_file.flush()",
+    "        os.fsync(tmp_file.fileno())",
+    "    if not is_v2:",
+    "        assert_legacy_source_stable(config_path, legacy_identity)",
+    "        if legacy_identity is None:",
+    "            os.link(tmp_name, config_path, follow_symlinks=False)",
+    "            os.unlink(tmp_name)",
+    "        else:",
+    "            os.replace(tmp_name, config_path)",
+    "    else:",
+    "        os.replace(tmp_name, config_path)",
+    "finally:",
+    "    try:",
+    "        os.unlink(tmp_name)",
+    "    except FileNotFoundError:",
+    "        pass",
+    "if is_v2:",
+    "    os.chmod(config_path, 0o600)",
+    "try:",
+    "    persisted = json.loads(config_path.read_text(encoding='utf-8')) if is_v2 else read_legacy_config(config_path)[0]",
+    "except (OSError, UnicodeDecodeError, ValueError) as exc:",
+    "    print(f'Could not verify managed MCP rollback state at {config_path}: {exc}', file=sys.stderr)",
+    "    raise SystemExit(2)",
+    "if is_v2:",
+    "    restored = persisted == {'mcpServers': payload['expectedServers']}",
+    "else:",
+    "    persisted_servers = persisted.get('mcpServers') if isinstance(persisted, dict) else None",
+    "    restored = isinstance(persisted_servers, dict) and persisted_servers.get(payload['server']) == payload['expected']",
+    "if not restored:",
+    "    print(f'Managed MCP rollback verification failed at {config_path}', file=sys.stderr)",
+    "    raise SystemExit(2)",
+    "print('NEMOCLAW_DEEPAGENTS_MCP_ROLLBACK_RESTORED=1')",
+    "PY",
+  ].join("\n");
+}
