+++ src/lib/actions/sandbox/mcp-bridge-adapter-deepagents-teardown.ts
@@ -1,0 +1,178 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import type { McpBridgeEntry } from "../../state/registry";
+import { runDeepAgentsAdapterCommand } from "./mcp-bridge-adapter-deepagents-command";
+import {
+  DEEPAGENTS_LEGACY_CONFIG_HELPERS,
+  DEEPAGENTS_LEGACY_MCP_CONFIG_PATH,
+} from "./mcp-bridge-adapter-deepagents-legacy";
+import type {
+  AdapterMutationOptions,
+  AdapterRemovalOutcome,
+} from "./mcp-bridge-adapter-inspection";
+import {
+  DEEPAGENTS_MCP_CONFIG_PATH,
+  deepAgentsManagedServerConfig,
+  pythonJsonLiteral,
+} from "./mcp-bridge-adapter-status";
+
+export function buildDeepAgentsMcpRemoveCommand(
+  entry: McpBridgeEntry,
+  force = false,
+  adaptiveTeardown = false,
+): string {
+  const payload = {
+    server: entry.server,
+    expected: deepAgentsManagedServerConfig(entry),
+    force,
+  };
+  return [
+    "/opt/venv/bin/python3 -I - <<'PY'",
+    "import json, os, pathlib, stat, sys, tempfile",
+    `payload = json.loads(${pythonJsonLiteral(payload)})`,
+    `managed_path = pathlib.Path(${JSON.stringify(DEEPAGENTS_MCP_CONFIG_PATH)})`,
+    `legacy_path = pathlib.Path(${JSON.stringify(DEEPAGENTS_LEGACY_MCP_CONFIG_PATH)})`,
+    ...DEEPAGENTS_LEGACY_CONFIG_HELPERS,
+    `runtime_kind = "${adaptiveTeardown ? "auto" : "v2"}"  # NEMOCLAW_DEEPAGENTS_RUNTIME_TEST_ANCHOR`,
+    "if runtime_kind == 'auto':",
+    "    runtime_kind = 'unknown'",
+    "    try:",
+    "        from deepagents_code import _nemoclaw_managed as managed",
+    "        runtime_path = str(getattr(managed, '_MCP_CONFIG_FILE', ''))",
+    "        if runtime_path == str(managed_path):",
+    "            runtime_kind = 'v2'",
+    "        elif runtime_path == str(legacy_path):",
+    "            runtime_kind = 'legacy'",
+    "    except Exception:",
+    "        pass",
+    "if runtime_kind not in ('v2', 'legacy'):",
+    "    print('Could not identify the managed Deep Agents MCP runtime; refusing teardown', file=sys.stderr)",
+    "    raise SystemExit(2)",
+    "is_v2 = runtime_kind == 'v2'",
+    "config_path = managed_path if is_v2 else legacy_path",
+    "legacy_identity = None",
+    "def finish(outcome):",
+    "    print('NEMOCLAW_DEEPAGENTS_MCP_REMOVAL=' + outcome)",
+    "    raise SystemExit(0)",
+    "def write_data(value):",
+    "    tmp_fd, tmp_name = tempfile.mkstemp(prefix='.nemoclaw-mcp.', dir=config_path.parent)",
+    "    try:",
+    "        os.fchmod(tmp_fd, 0o600)",
+    "        with os.fdopen(tmp_fd, 'w', encoding='utf-8') as tmp_file:",
+    "            json.dump(value, tmp_file, indent=2, sort_keys=True)",
+    "            tmp_file.write('\\n')",
+    "            tmp_file.flush()",
+    "            os.fsync(tmp_file.fileno())",
+    "        if not is_v2:",
+    "            assert_legacy_source_stable(config_path, legacy_identity)",
+    "            if legacy_identity is None:",
+    "                os.link(tmp_name, config_path, follow_symlinks=False)",
+    "                os.unlink(tmp_name)",
+    "            else:",
+    "                os.replace(tmp_name, config_path)",
+    "        else:",
+    "            os.replace(tmp_name, config_path)",
+    "    finally:",
+    "        try:",
+    "            os.unlink(tmp_name)",
+    "        except FileNotFoundError:",
+    "            pass",
+    "    if is_v2:",
+    "        os.chmod(config_path, 0o600)",
+    "if not os.path.lexists(config_path):",
+    "    finish('absent')",
+    "if config_path.is_symlink():",
+    "    if is_v2 and payload['force']:",
+    "        config_path.unlink()",
+    "        finish('removed')",
+    "    if is_v2:",
+    "        print(f'Invalid managed MCP v2 projection symlink at {config_path}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    finish('unowned')",
+    "try:",
+    "    if is_v2:",
+    "        data = json.loads(config_path.read_text(encoding='utf-8') or '{}')",
+    "    else:",
+    "        data, legacy_identity = read_legacy_config(config_path)",
+    "except (OSError, UnicodeDecodeError, ValueError) as exc:",
+    "    if is_v2 and payload['force']:",
+    "        config_path.unlink()",
+    "        finish('removed')",
+    "    if is_v2:",
+    "        print(f'Invalid managed MCP v2 projection at {config_path}: {exc}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    finish('unowned')",
+    "if not isinstance(data, dict):",
+    "    if is_v2 and payload['force']:",
+    "        config_path.unlink()",
+    "        finish('removed')",
+    "    if is_v2:",
+    "        print(f'Invalid managed MCP v2 projection at {config_path}: expected object', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    finish('unowned')",
+    "servers = data.get('mcpServers')",
+    "if not isinstance(servers, dict):",
+    "    if not is_v2 and 'mcpServers' not in data:",
+    "        finish('absent')",
+    "    if is_v2 and payload['force']:",
+    "        config_path.unlink()",
+    "        finish('removed')",
+    "    if is_v2:",
+    "        print(f'Invalid managed MCP v2 server map at {config_path}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    finish('unowned')",
+    "present = payload['server'] in servers",
+    "current = servers.get(payload['server'])",
+    "if is_v2:",
+    "    if data and set(data) != {'mcpServers'}:",
+    "        if payload['force']:",
+    "            config_path.unlink()",
+    "            finish('removed')",
+    "        print(f'Invalid managed MCP v2 projection at {config_path}: only mcpServers is allowed', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    if present and not payload['force'] and current != payload['expected']:",
+    "        print(f\"Refusing to remove modified MCP server '{payload['server']}' from {config_path}. Use --force to remove it.\", file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "    if not present:",
+    "        finish('absent')",
+    "else:",
+    "    if not present:",
+    "        finish('absent')",
+    "    if current != payload['expected'] and not payload['force']:",
+    "        finish('unowned')",
+    "servers.pop(payload['server'])",
+    "if not servers:",
+    "    data.pop('mcpServers', None)",
+    "if data:",
+    "    write_data(data)",
+    "    if not is_v2 and read_legacy_config(config_path)[0] != data:",
+    "        print(f'Legacy MCP teardown verification failed at {config_path}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "else:",
+    "    if not is_v2:",
+    "        assert_legacy_source_stable(config_path, legacy_identity)",
+    "    config_path.unlink()",
+    "    if os.path.lexists(config_path):",
+    "        print(f'Managed MCP teardown verification failed at {config_path}', file=sys.stderr)",
+    "        raise SystemExit(2)",
+    "finish('removed')",
+    "PY",
+  ].join("\n");
+}
+
+export function unregisterDeepAgentsAdapter(
+  sandboxName: string,
+  entry: McpBridgeEntry,
+  options: AdapterMutationOptions = {},
+): AdapterRemovalOutcome {
+  const stdout = runDeepAgentsAdapterCommand(
+    sandboxName,
+    entry,
+    buildDeepAgentsMcpRemoveCommand(entry, options.force === true, options.teardown === true),
+    `Deep Agents Code MCP config removal failed for '${entry.server}'.`,
+    options,
+  );
+  const marker = stdout.match(/NEMOCLAW_DEEPAGENTS_MCP_REMOVAL=(removed|absent|unowned)/);
+  return (marker?.[1] as AdapterRemovalOutcome | undefined) ?? "unowned";
+}
